fixed several similar instances of a bug that would result in a crash

+++ VVOSC/FrameworkSrc/OSCValue.m
@@ -48,7 +48,7 @@
 }
 - (NSString *) lengthyDescription	{
 #if !IPHONE
-	float		colorComps[4];
+	CGFloat		colorComps[4];
 #endif
 	switch (type)	{
 		case OSCValInt:
